requirements.txt
pandas
sqlalchemy
psycopg2
apache-airflow
matplotlib
seaborn